backend/requirements.txt
# Web Framework
fastapi>=0.115.5
uvicorn[standard]>=0.32.1

# AI & RAG
anthropic>=0.40.0
chromadb>=0.6.0
sentence-transformers>=3.3.1

# Document Processing
pymupdf>=1.24.14
pytesseract>=0.3.13
Pillow>=11.0.0
python-multipart>=0.0.12
pdfplumber>=0.11.4

# Database
supabase>=2.10.0

# Utilities
python-dotenv>=1.0.1
pydantic>=2.11.0
pydantic-settings>=2.7.0
httpx>=0.27.2
aiofiles>=24.1.0
